ADD - claims to JWT for admin privileges

+++ controllers/item.py
@@ -3,7 +3,7 @@ from typing import Dict, Any
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from controllers.utils import ItemNotFoundError, handle_error
 from models import ItemModel
@@ -23,6 +23,10 @@ class Item(MethodView):
 
     @jwt_required()
     def delete(self, item_id: int):
+        jwt = get_jwt()
+        if not jwt.get("is_admin", False):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
